Updated remoteID of photos on collection rename
The collection name is part of the photo of remote ID, so that when the
collection is renamed, the remote ID of each photo in that collection must also
be updated. The reason for using paths in remote IDs is that it is the only way
to address the files/dirs on the remote service.